requirements.txt
fastapi==0.111.0
pymongo==4.7.2
motor==3.4.0
openai==1.30.5
pandas==2.2.2
pika==1.3.2
requests==2.32.3
regex==2024.5.15
matplotlib==3.9.0
tensorflow==2.16.1
PyMuPDF==1.24.5
python-socketio==5.11.2
aio-pika==9.4.1
aiofiles==23.1.0
aiobotocore==2.13.0
pycurl==7.45.3
google-cloud-storage==2.18.0
python-jose==3.3.0
